fix: address code review findings
- HTMLWebView: fix `result as? CGFloat` always failing (JS returns
  Double, not CGFloat) — HTML emails were stuck at 100pt height
- DevSeeder: add system labels and ThreadLabelRecord rows so DEBUG
  inbox isn't empty with label-based filtering

+++ Packages/Features/ThreadFeature/Sources/ThreadFeature/HTMLWebView.swift
@@ -103,9 +103,9 @@ struct HTMLWebView: NSViewRepresentable {
 
         func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
             webView.evaluateJavaScript("document.body.scrollHeight") { [weak self] result, _ in
-                if let height = result as? CGFloat {
+                if let height = result as? Double {
                     Task { @MainActor in
-                        self?.heightBinding?.wrappedValue = min(height, 2000)
+                        self?.heightBinding?.wrappedValue = min(CGFloat(height), 2000)
                     }
                 }
             }
